+++ integration-test/entrypoint.clj
@@ -7,7 +7,8 @@
 (def namespaces
   '[integration.initialize-test
     integration.chat.openai-test
-    integration.chat.anthropic-test])
+    integration.chat.anthropic-test
+    integration.chat.ollama-test])
 
 (defn timeout [timeout-ms callback]
   (let [fut (future (callback))